Moving towards a more general filter system
Implements an image processing framework supporting a single filter (currently only the LUT filter is implemented). This framework switches to using floating point color because it's one of the design goals for the system and I wanted to establish at least baseline performance before it gets more complicated.

Filter strength and precalculation are now baked into the filter and are not LUT-specific.