Moving towards composing filters dynamically
Adds a new Look type which contains a sequence of filters to apply. Filters are applied by accumulating into a single 3D LUT which is applied to the input image in one go.

This commit also starts to build-out a sort of resource system for managing data loaded from disk (e.g. images, LUTs). This will eventually be used for saving/loading compositions to/from disk.