Adds a mask preview overlay to show mask shape
Adds a new option into the "View" menu which enables an overlay in the canvas which shows the mask on top of the image. The mask is red at maximum value, and transparent at minimum value.

The mask overlay image is generated by an OpenCL kernel in a separate processor.