Add initial version of dynamic data API
This is intended to eventually power more dynamic IO and UI features, making it easier to implement new filters.

Makes heavy usage of constexpr, static initialization and macro magic.

In this version, data types and properties are registered with metadata for runtime look-up. There is also support for interfaces (of a sort). Lists of properties are not supported.